requirements.txt
pathlib
pyyaml
fastapi
uvicorn